pyproject.toml
[tool.poetry]
name = "stand-up-comedy"
version = "0.1.0"
description = ""
authors = ["Pau <plabartabajo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
jupyter = "^1.0.0"
youtube-transcript-api = "^0.4.1"
selenium = "^3.141.0"
transformers = "^4.5.1"
torch = "^1.8.1"
pandas = "^1.2.4"
nltk = "^3.6.2"
seaborn = "^0.11.1"
pydrive = "^1.3.1"
pytorch_lightning = "^1.2.10"
python-dotenv = "^0.17.0"
datasets = "^1.6.2"
sklearn = "^0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
